+++ com.unity.formats.alembic/Runtime/Scripts/Exporter/AbcAPI.cs
@@ -355,7 +355,7 @@ namespace UnityEngine.Formats.Alembic.Sdk
 
 
         [DllImport(Abci.Lib, BestFitMapping = false, ThrowOnUnmappableChar = true)] public static extern void aiClearContextsWithPath(string path);
-        [DllImport(Abci.Lib)] public static extern aiContext aiContextCreate(int uid);
+        [DllImport(Abci.Lib)] public static extern aiContext aiContextCreate(UInt64 uid);
         [DllImport(Abci.Lib)] public static extern void aiContextDestroy(IntPtr ctx);
         [DllImport(Abci.Lib, BestFitMapping = false, ThrowOnUnmappableChar = true)] public static extern Bool aiContextLoad(IntPtr ctx, string path);
         [DllImport(Abci.Lib)] public static extern void aiContextSetConfig(IntPtr ctx, ref aiConfig conf);
